Fix string overflow bug
We used ';' as a second parameter to the string constructor, which is clearly wrong and made the constructor read 59 bytes instead of using the null byte as termination.

https://en.cppreference.com/w/cpp/string/basic_string/basic_string

Not that it'd be a big deal security wise (the value has never been printed, so this couldn't be used to e.g., read secrets from RAM in CI environments etc.), but it breaks the functionality.

Found while looking into https://github.com/linuxdeploy/linuxdeploy-plugin-qt/issues/61.