Move build in Docker into scripts
Debugging a build on GitHub actions is very time consuming, boring and
therefore very annoying if it takes more than a couple of runs. The new
scripts allow running the builds on any kind of infrastructure that
supports Docker, and minimize the coupling to GitHub actions overall.

A few fixes were also needed in the build infra to fix the compilation
on Debian stable.